+++ hadoop-tools/hadoop-azure/src/main/java/org/apache/hadoop/fs/azurebfs/constants/FileSystemConfigurations.java
@@ -114,7 +114,7 @@ public final class FileSystemConfigurations {
   public static final String DEFAULT_VALUE_UNKNOWN = "UNKNOWN";
 
   public static final boolean DEFAULT_DELETE_CONSIDERED_IDEMPOTENT = true;
-  public static final boolean DEFAULT_ENABLE_PAGINATED_DELETE = true;
+  public static final boolean DEFAULT_ENABLE_PAGINATED_DELETE = false;
   public static final int DEFAULT_CLOCK_SKEW_WITH_SERVER_IN_MS = 5 * 60 * 1000; // 5 mins
 
   public static final int STREAM_ID_LEN = 12;
